change path /books to /library, fix css position for pagination
