+++ app/(root)/layout.tsx
@@ -33,7 +33,7 @@ const layout = async ({ children }: { children: ReactNode }) => {
     <main className="root-container">
       <div className="mx-auto max-w-7xl">
         <Header session={session} />
-        <div className="mt-20 pb-20">{children}</div>
+        <div className="mt-20">{children}</div>
       </div>
     </main>
   );
